benchmarks/requirements_bench.txt
langchain-text-splitters
matplotlib
google-generativeai
python-dotenv
